Encrypted buffers are bigger (xxtea adds a few bytes to the encrypted buffer so the) Fix byte reading (use the offset from the stream channel)

+++ src/com/philip1337/veloxio/utils/Base64.java
@@ -92,7 +92,7 @@ final class Base64 {
             if (b2 == -1) {
                 break;
             }
-            buf.write((int) ((b1 << 2) | ((b2 & 0x30) >>> 4)));
+            buf.write((b1 << 2) | ((b2 & 0x30) >>> 4));
 
             /* b3 */
             do {
@@ -105,7 +105,7 @@ final class Base64 {
             if (b3 == -1) {
                 break;
             }
-            buf.write((int) (((b2 & 0x0f) << 4) | ((b3 & 0x3c) >>> 2)));
+            buf.write(((b2 & 0x0f) << 4) | ((b3 & 0x3c) >>> 2));
 
             /* b4 */
             do {
@@ -118,7 +118,7 @@ final class Base64 {
             if (b4 == -1) {
                 break;
             }
-            buf.write((int) (((b3 & 0x03) << 6) | b4));
+            buf.write(((b3 & 0x03) << 6) | b4);
         }
         return buf.toByteArray();
     }
